Move Tweet serching logic to separate file
